RCDY-308 chore: changes from moving category events to core
