+++ service-catalog/content/impl/src/main/java/com/recody/recodybackend/catalog/features/category/delete/DeleteCategoryHandler.java
@@ -1,6 +1,6 @@
 package com.recody.recodybackend.catalog.features.category.delete;
 
-import com.recody.recodybackend.common.events.CategoryDeleted;
+import com.recody.recodybackend.event.CategoryDeleted;
 
 public interface DeleteCategoryHandler {
