+++ service-catalog/content/impl/src/main/java/com/recody/recodybackend/catalog/features/changecategoryoncontent/ChangeCategoryOnContentHandler.java
@@ -1,6 +1,6 @@
 package com.recody.recodybackend.catalog.features.changecategoryoncontent;
 
-import com.recody.recodybackend.common.events.CategoryPersonalized;
+import com.recody.recodybackend.event.CategoryPersonalized;
 
 public interface ChangeCategoryOnContentHandler {
